Move action_context into separate file
