+++ tests/unit_tests/tags/action_context.cpp
@@ -1,6 +1,7 @@
+#include "tags/action_context.hpp"
+
 #include "common/test.hpp"
 #include "components/logger.hpp"
-#include "tags/context.hpp"
 
 using namespace polybar;
 using namespace std;
